+++ gridopt/power_flow/ac_pf.py
@@ -22,6 +22,7 @@ class ACPF(PFmethod):
     CONTROL_MODE_LOCKED = 'locked'
     CONTROL_MODE_FREE = 'free'
     CONTROL_MODE_REG = 'regulating'
+    CONTROL_MODE_DISC = 'discrete'
 
     name = 'ACPF'
 
@@ -33,13 +34,15 @@ class ACPF(PFmethod):
                    'weight_redispatch': 1e0, # weight for gen real power redispatch deviation penalty
                    'v_min_clip': 0.5,        # lower v threshold for clipping
                    'v_max_clip': 1.5,        # upper v threshold for clipping
-                   'v_limits': False,        # voltage magnitude limits
+                   'v_limits': False,        # voltage magnitude limits (OPT only)
                    'Q_limits': True,         # flag for enforcing generator, VSC and FACTS reactive power limits
                    'Q_mode': 'regulating',   # reactive power mode: free, regulating
                    'shunt_limits': True,     # flag for enforcing switched shunt susceptance limits
                    'shunt_mode': 'locked',   # switched shunts mode: locked, free, regulating
                    'tap_limits': True,       # flag for enforcing transformer tap ratio limits
                    'tap_mode': 'locked',     # transformer tap ratio mode: locked, free, regulating
+                   'phase_shift_limits': True,      # flag for enforcing transformer tap ratio limits
+                   'phase_shift_mode': 'locked',    # transformer phase-shift mode: locked, free, regulating
                    'lock_vsc_P_dc': True,    # flag for locking vsc P dc
                    'lock_csc_P_dc': True,    # flag for locking csc P dc
                    'lock_csc_i_dc': True,    # flag for locking csc i dc
@@ -47,27 +50,32 @@ class ACPF(PFmethod):
                    'pvpq_start_k': 0,        # start iteration number for PVPQ switching heuristics
                    'vmin_thresh': 0.1,       # minimum voltage magnitude threshold
                    'gens_redispatch': False, # flag for allowing active power redispatch
-                   'shunts_round': True,     # flag for rounding discrete switched shunt susceptances (not supported yet)
-                   'taps_round': True,       # flag for rounding discrete transformer tap ratios (not supported yet)
+                #    'shunts_round': False,     # flag for rounding discrete switched shunt susceptances (not supported yet)
+                   'taps_round': False,      # flag for rounding discrete transformer tap ratios (NR only)
                    'v_mag_warm_ref': False,  # flag for using current v_mag as reference in v_mag regularization
                    'solver': 'nr',           # OPTALG optimization solver: augl, ipopt, nr, inlp
                    'tap_step': 0.5,          # tap ratio acceleration factor (NR only)
                    'shunt_step': 0.5,        # susceptance acceleration factor (NR only)
                    'dtap': 1e-4,             # tap ratio perturbation (NR only)
+                   'dsus': 1e-4,             # susceptance perturbation (NR only)
+                   'y_correction': True,     # admittance correction (NR only)
                    'load_q_curtail': False,  # flag for allowing load Q to change (OPT only)
                    'P_transfer': False,      # flag for enabling inter-area transfer or not
-                   'dsus': 1e-4}             # susceptance perturbation (NR only)
+                   'dsus': 1e-4,             # susceptance perturbation (NR only)
+                   'v_min_2_ZIP': 0.85,      # minimum voltage threshold to convert loads to constant impedance
+                   'loads_2_ZIP': True}      # Flag to convert loads to constant impedance if the voltage drops below
+                                             # v_min_2_ZIP
 
-    _parameters_augl = {'feastol' : 1e-4,
-                        'optol' : 1e0,
-                        'kappa' : 1e-5,
+    _parameters_augl = {'feastol': 1e-4,
+                        'optol': 1e0,
+                        'kappa': 1e-5,
                         'theta_max': 1e-6,
                         'sigma_init_max': 1e9}
 
     _parameters_ipopt = {}
 
-    _parameters_inlp = {'feastol' : 1e-4,
-                        'optol' : 1e0}
+    _parameters_inlp = {'feastol': 1e-4,
+                        'optol': 1e0}
 
     _parameters_nr = {}
 
@@ -118,26 +126,37 @@ class ACPF(PFmethod):
         shunt_limits = params['shunt_limits']
         tap_mode = params['tap_mode']
         tap_limits = params['tap_limits']
+        phase_shift_mode = params['phase_shift_mode']
+        phase_shift_limits = params['phase_shift_limits']
         lock_vsc_P_dc = params['lock_vsc_P_dc']
         lock_csc_P_dc = params['lock_csc_P_dc']
         lock_csc_i_dc = params['lock_csc_i_dc']
         vdep_loads = params['vdep_loads']
         gens_redispatch = params['gens_redispatch']
         P_transfer = params['P_transfer']
+        convert_loads_2_zip = params['loads_2_ZIP']
+        y_correction = params['y_correction']
 
         # Check shunt options
         if shunt_mode not in [self.CONTROL_MODE_LOCKED,
-                               self.CONTROL_MODE_REG]:
+                              self.CONTROL_MODE_REG]:
             raise ValueError('invalid shunts mode')
         if shunt_mode == self.CONTROL_MODE_REG and not shunt_limits:
             raise ValueError('unsupported shunts configuration')
 
         # Check tap options
         if tap_mode not in [self.CONTROL_MODE_LOCKED,
-                               self.CONTROL_MODE_REG]:
+                            self.CONTROL_MODE_REG]:
             raise ValueError('invalid taps mode')
         if tap_mode == self.CONTROL_MODE_REG and not tap_limits:
             raise ValueError('unsupported taps configuration')
+
+        # Check phase-shift options
+        if phase_shift_mode not in [self.CONTROL_MODE_LOCKED,
+                                    self.CONTROL_MODE_REG]:
+            raise ValueError('invalid phase-shifter mode')
+        if phase_shift_mode == self.CONTROL_MODE_REG and not phase_shift_limits:
+            raise ValueError('unsupported phase-shifter configuration')
 
         # Check Q options
         if Q_mode != self.CONTROL_MODE_REG:
@@ -177,6 +196,11 @@ class ACPF(PFmethod):
                       'regulator',
                       'reactive power')
 
+        net.set_flags('generator',
+                      'variable',
+                      'fixed power factor',
+                      'reactive power')
+
         # VSC HVDC
         net.set_flags('vsc converter',
                       'variable',
@@ -201,20 +225,46 @@ class ACPF(PFmethod):
                       'any',
                       'all')
 
-        # Loads
-        if vdep_loads:
+        # Load
+        if vdep_loads or convert_loads_2_zip:
             for load in net.loads:
-                if load.is_voltage_dependent() and load.is_in_service():
-                    net.set_flags_of_component(load,
-                                               'variable',
-                                               ['active power', 'reactive power'])
+                if load.is_in_service():
+                    if convert_loads_2_zip:      # Consider all loads
+                        net.set_flags_of_component(load,
+                                                   'variable',
+                                                   ['active power', 'reactive power'])
+                    if load.is_voltage_dependent():  # Consider only voltage dependent loads
+                        net.set_flags_of_component(load,
+                                                   'variable',
+                                                   ['active power', 'reactive power'])
 
-        # Tap changers
+        # Tap changer
         if tap_mode != self.CONTROL_MODE_LOCKED:
             net.set_flags('branch',
                           ['variable', 'fixed'],
                           'tap changer - v',
                           'tap ratio')
+            net.set_flags('branch',
+                          'variable',
+                          'tap changer - Q',
+                          'tap ratio')
+            if y_correction:
+                net.set_flags('branch', 
+                              'variable',
+                              'y correction - ratio', 
+                              'y scale') 
+        
+        # Phase-shifters
+        if phase_shift_mode != self.CONTROL_MODE_LOCKED:
+            net.set_flags('branch',
+                          'variable',
+                          'phase shifter',
+                          'phase shift')
+            if y_correction:
+                net.set_flags('branch', 
+                              'variable',
+                              'y correction - phase',
+                              'y scale')
 
         # Switched shunts
         if shunt_mode != self.CONTROL_MODE_LOCKED:
@@ -245,13 +295,16 @@ class ACPF(PFmethod):
         problem.add_constraint(pfnet.Constraint('VSC DC power control', net))
         problem.add_constraint(pfnet.Constraint('CSC DC power control', net))
         problem.add_constraint(pfnet.Constraint('CSC DC current control', net))
+        problem.add_constraint(pfnet.Constraint('fixed gen constant power factor', net))
 
         problem.add_constraint(pfnet.Constraint('PVPQ switching', net))
         problem.add_constraint(pfnet.Constraint('switching power factor regulation', net))
         problem.add_constraint(pfnet.Constraint('switching FACTS active power control', net))
         problem.add_constraint(pfnet.Constraint('switching FACTS reactive power control', net))
+        problem.add_constraint(pfnet.Constraint('switching FACTS series voltage control', net))
+        problem.add_constraint(pfnet.Constraint('switching FACTS series impedance control', net))
 
-        if vdep_loads:
+        if vdep_loads or convert_loads_2_zip:
             problem.add_constraint(pfnet.Constraint('load voltage dependence', net))
 
         if Q_limits:
@@ -261,6 +314,17 @@ class ACPF(PFmethod):
         if P_transfer:
             problem.add_constraint(pfnet.Constraint('switching area transfer regulation', net))
             problem.add_heuristic(pfnet.Heuristic('switching area transfer regulation', net))
+        
+        if tap_mode != self.CONTROL_MODE_LOCKED:
+            problem.add_constraint(pfnet.Constraint('switching transformer q regulation', net))
+            problem.add_heuristic(pfnet.Heuristic('switching transformer q regulation', net))
+
+        if phase_shift_mode != self.CONTROL_MODE_LOCKED:
+            problem.add_constraint(pfnet.Constraint('switching transformer p regulation', net))
+            problem.add_heuristic(pfnet.Heuristic('switching transformer p regulation', net))
+
+        if y_correction and (tap_mode != self.CONTROL_MODE_LOCKED or phase_shift_mode != self.CONTROL_MODE_LOCKED):
+            problem.add_constraint(pfnet.Constraint('y correction table', net))
 
         problem.analyze()
 
@@ -290,6 +354,8 @@ class ACPF(PFmethod):
         shunt_limits = params['shunt_limits']
         tap_mode = params['tap_mode']
         tap_limits = params['tap_limits']
+        phase_shift_mode = params['phase_shift_mode']
+        phase_shift_limits = params['phase_shift_limits']
         lock_vsc_P_dc = params['lock_vsc_P_dc']
         lock_csc_P_dc = params['lock_csc_P_dc']
         lock_csc_i_dc = params['lock_csc_i_dc']
@@ -298,22 +364,31 @@ class ACPF(PFmethod):
         gens_redispatch = params['gens_redispatch']
         curtail_load_q = params['load_q_curtail']
         P_transfer = params['P_transfer']
+        y_correction = params['y_correction']
 
         # Check shunt options
         if shunt_mode not in [self.CONTROL_MODE_LOCKED,
-                               self.CONTROL_MODE_FREE,
-                               self.CONTROL_MODE_REG]:
+                              self.CONTROL_MODE_FREE,
+                              self.CONTROL_MODE_REG]:
             raise ValueError('invalid shunts mode')
         if shunt_mode == self.CONTROL_MODE_REG and not shunt_limits:
             raise ValueError('unsupported shunts configuration')
 
         # Check tap options
         if tap_mode not in [self.CONTROL_MODE_LOCKED,
-                             self.CONTROL_MODE_FREE,
-                             self.CONTROL_MODE_REG]:
+                            self.CONTROL_MODE_FREE,
+                            self.CONTROL_MODE_REG]:
             raise ValueError('invalid taps mode')
         if tap_mode == self.CONTROL_MODE_REG and not tap_limits:
             raise ValueError('unsupported taps configuration')
+
+        # Check tap options
+        if phase_shift_mode not in [self.CONTROL_MODE_LOCKED,
+                                    self.CONTROL_MODE_FREE,
+                                    self.CONTROL_MODE_REG]:
+            raise ValueError('invalid phase-shifter mode')
+        if phase_shift_mode == self.CONTROL_MODE_REG and not phase_shift_limits:
+            raise ValueError('unsupported phase-shifter configuration')
 
         # Check Q options
         if Q_mode not in [self.CONTROL_MODE_REG,
@@ -343,7 +418,7 @@ class ACPF(PFmethod):
                           'any',
                           'voltage magnitude')
 
-        # Genertors
+        # Generators
         if gens_redispatch:
             # Assume slack gens (excep renewables) are redispatchable
             net.set_flags('generator',
@@ -358,6 +433,10 @@ class ACPF(PFmethod):
         net.set_flags('generator',
                       'variable',
                       'regulator',
+                      'reactive power')
+        net.set_flags('generator',
+                      'variable',
+                      'fixed power factor',
                       'reactive power')
         if Q_mode == self.CONTROL_MODE_FREE and Q_limits:
             net.set_flags('generator',
@@ -414,19 +493,58 @@ class ACPF(PFmethod):
                           'any',
                           'reactive power')
 
-        # Tap changers
+        # Trans tap mode
         if tap_mode != self.CONTROL_MODE_LOCKED:
             net.set_flags('branch',
                           'variable',
                           'tap changer - v',
                           'tap ratio')
-        if tap_mode == self.CONTROL_MODE_FREE and tap_limits:
             net.set_flags('branch',
-                          'bounded',
-                          'tap changer - v',
+                          'variable',
+                          'tap changer - Q',
                           'tap ratio')
+            # Trans reg limit
+            if tap_limits:
+                net.set_flags('branch',
+                            'bounded',
+                            'tap changer - v',
+                            'tap ratio')
+                net.set_flags('branch',
+                            'bounded',
+                            'tap changer - Q',
+                            'tap ratio')
+            # Y correction
+            if y_correction:
+                net.set_flags('branch',
+                              'variable',
+                              'y correction - ratio',
+                              ['tap ratio', 'y scale'])
 
-        # Swtiched shunts
+        # Trans phase-shifter mode
+        if phase_shift_mode != self.CONTROL_MODE_LOCKED:
+            net.set_flags('branch',
+                          'variable',
+                          'phase shifter',
+                          'phase shift')
+            # Reg limit
+            if phase_shift_limits:
+                net.set_flags('branch',
+                              'bounded',
+                              'phase shifter',
+                              'phase shift')
+                if net.get_num_asymmetric_phase_shifters() > 0:
+                    net.set_flags('branch',
+                                  ['variable', 'bounded'],
+                                  'asymmetric phase shifter',
+                                  'ratio')
+            # Y correction
+            if y_correction:
+                net.set_flags('branch',
+                              'variable',
+                              'y correction - phase',
+                              ['phase shift', 'y scale'])
+
+        # Switched shunts
         if shunt_mode != self.CONTROL_MODE_LOCKED:
             net.set_flags('shunt',
                           'variable',
@@ -463,6 +581,9 @@ class ACPF(PFmethod):
         problem.add_constraint(pfnet.Constraint('VSC DC voltage control', net))
         problem.add_constraint(pfnet.Constraint('CSC DC voltage control', net))
         problem.add_constraint(pfnet.Constraint('power factor regulation', net))
+        problem.add_constraint(pfnet.Constraint('switching FACTS series voltage control', net))
+        problem.add_constraint(pfnet.Constraint('switching FACTS series impedance control', net))
+        problem.add_constraint(pfnet.Constraint('fixed gen constant power factor', net))
 
         if lock_vsc_P_dc:
             problem.add_constraint(pfnet.Constraint('VSC DC power control', net))
@@ -494,9 +615,20 @@ class ACPF(PFmethod):
             problem.add_constraint(pfnet.Constraint('variable fixing', net))
 
         if tap_mode != self.CONTROL_MODE_LOCKED:
-            problem.add_function(pfnet.Function('tap ratio regularization', wc/(net.get_num_tap_changers_v(True)+1.), net))
+            problem.add_function(pfnet.Function('tap ratio regularization', wc/(net.get_num_tap_changers(True)+1.), net))
             if tap_mode == self.CONTROL_MODE_REG and tap_limits:
+                problem.add_function(pfnet.Function('transformer Q regularization', wc/(net.get_num_tap_changers_Q(True)+1.), net))
                 problem.add_constraint(pfnet.Constraint('voltage regulation by transformers', net))
+
+        if phase_shift_mode != self.CONTROL_MODE_LOCKED:
+            problem.add_function(pfnet.Function('phase shift regularization', wc/(net.get_num_phase_shifters(True)+1.), net))
+            if phase_shift_mode == self.CONTROL_MODE_REG and phase_shift_limits:
+                problem.add_function(pfnet.Function('transformer P regularization', wc/(net.get_num_phase_shifters(True)+1.), net))
+                if net.get_num_asymmetric_phase_shifters() > 0:
+                    problem.add_constraint(pfnet.Constraint('asymmetric transformer equations', net))
+
+        if y_correction and (tap_mode != self.CONTROL_MODE_LOCKED or phase_shift_mode != self.CONTROL_MODE_LOCKED):
+            problem.add_constraint(pfnet.Constraint("y correction table", net))
 
         if shunt_mode != self.CONTROL_MODE_LOCKED:
             problem.add_function(pfnet.Function('susceptance regularization', wc/(net.get_num_switched_v_shunts(True)+1.), net))
@@ -556,6 +688,9 @@ class ACPF(PFmethod):
         vmin_thresh = params['vmin_thresh']
         solver_name = params['solver']
         solver_params = params['solver_parameters']
+        quiet = solver_params[solver_name]['quiet']
+        convert_loads_2_zip = params['loads_2_ZIP']
+        vmin_to_zip = params['v_min_2_ZIP']
 
         # Opt solver
         if solver_name == 'augl':
@@ -572,14 +707,18 @@ class ACPF(PFmethod):
 
         # Callbacks
         def c1(s):
+            """Apply Transformer Voltage Regulation"""
             if (s.k != 0 and params['tap_limits'] and tap_mode == self.CONTROL_MODE_REG and
                 norm(s.problem.f, np.inf) < 100.*solver_params['nr']['feastol']):
                 try:
                     self.apply_tran_v_regulation(s)
+                    if params['taps_round']:
+                        self.apply_tran_tap_rounding(s)
                 except Exception as e:
                     raise PFmethodError_TranVReg(e)
-
+                
         def c2(s):
+            """Apply Shunt Voltage Regulation"""
             if (s.k != 0 and params['shunt_limits'] and shunt_mode == self.CONTROL_MODE_REG and
                 norm(s.problem.f, np.inf) < 100.*solver_params['nr']['feastol']):
                 try:
@@ -588,16 +727,35 @@ class ACPF(PFmethod):
                     raise PFmethodError_ShuntVReg(e)
 
         def c3(s):
+            """PV-PQ start at given iteration"""
             if (s.k >= params['pvpq_start_k'] and params['Q_limits'] and Q_mode == self.CONTROL_MODE_REG):
                 prob = s.problem.wrapped_problem
                 prob.apply_heuristics(s.x)
                 s.problem.A = prob.A
                 s.problem.b = prob.b
 
+        def c4(s):
+            net = s.problem.wrapped_problem.network
+            if np.min(s.problem.wrapped_problem.network.bus_v_min) <= vmin_to_zip:
+                for lod in net.loads:
+                    if lod.is_in_service():
+                        bus_v = s.get_primal_variables()[lod.bus.index_v_mag]
+                        if bus_v <= vmin_to_zip:
+                            # Move cp&ci to cg
+                            lod.comp_cg = lod.comp_cp + lod.comp_ci + lod.comp_cg
+                            lod.comp_cp = 0.0
+                            lod.comp_ci = 0.0
+                            # Move cq&cj to cb
+                            lod.comp_cb = lod.comp_cb - lod.comp_cq - lod.comp_cj
+                            lod.comp_cq = 0.0
+                            lod.comp_cj = 0.0
+
         if solver_name == 'nr':
             solver.add_callback(OptCallback(c1))
             solver.add_callback(OptCallback(c2))
             solver.add_callback(OptCallback(c3))
+            if convert_loads_2_zip:
+                solver.add_callback(OptCallback(c4))
 
         # Termination
         def t1(s):
@@ -633,6 +791,13 @@ class ACPF(PFmethod):
                     net.clear_sensitivities()
                     if solver_name != 'nr':
                         problem.store_sensitivities(*solver.get_dual_variables())
+
+            if not quiet and convert_loads_2_zip and solver.problem.wrapped_problem.network.bus_v_min <= vmin_to_zip:
+                for lod in net.loads:
+                    if lod.is_in_service():
+                        bus_v = solver.get_primal_variables()[lod.bus.index_v_mag]
+                        if bus_v <= vmin_to_zip:
+                            print(f"  Converted load {lod.name}, at bus {lod.bus.number} with v_mag={bus_v:>1.4f} to ZIP load")
 
             # Save results
             self.set_solver_name(solver_name)
@@ -807,7 +972,6 @@ class ACPF(PFmethod):
             if bus.is_regulated_by_tran(True) and not bus.is_slack():
 
                 assert(bus.has_flags('variable','voltage magnitude'))
-
                 for tau in range(net.num_periods):
 
                     v = x[bus.index_v_mag[tau]]
@@ -855,6 +1019,71 @@ class ACPF(PFmethod):
                             if np.abs(tnew-t) > eps:
                                 break
 
+        # Update
+        solver.func(x)
+        p.update_lin()
+        solver.problem.A = p.A
+        solver.problem.b = p.b
+
+    def apply_tran_tap_rounding(self, solver):
+        """Apply transformer voltage regualtion using descrete mode."""
+
+        # Local variables
+        p = solver.problem.wrapped_problem
+        net = p.network
+        x = solver.x
+        eps = 1e-8
+
+        # Fix constraints
+        c = p.find_constraint('variable fixing')
+        A = c.A
+        b = c.b
+
+        for i in range(net.num_branches):
+
+            br = net.get_branch(i)
+
+            if br.is_in_service() and br.is_tap_changer():
+
+                assert(br.reg_bus.has_flags('variable', 'voltage magnitude'))
+                for tau in range(net.num_periods):
+
+                    if(br.has_flags('variable', 'tap ratio')):
+                        # Ratio is variable
+                        cur_ratio = x[br.index_ratio[tau]]
+                        if cur_ratio > br.ratio_max:
+                            x[br.index_ratio[tau]] = br.ratio_max
+                        elif cur_ratio < br.ratio_min:
+                            x[br.index_ratio[tau]] = br.ratio_min
+                        else:
+                            # Find closest ratio
+                            dratio = (br.ratio_max - br.ratio_min) / (br.num_ratios-1)
+                            ptn_ratios = np.arange(0, br.num_ratios)*dratio + br.ratio_min
+                            i_close = np.argmin(np.abs(ptn_ratios - cur_ratio))
+                            new_ratio = ptn_ratios[i_close]
+
+                            # Fix constr index
+                            k = int(np.where(A.col == br.index_ratio[tau])[0])
+                            ik = A.row[k]
+                            assert(np.abs(b[ik]-x[br.index_ratio[tau]]) < eps)
+                            assert(A.data[k] == 1.)
+
+                            # Update var and constraint
+                            x[br.index_ratio[tau]] = new_ratio
+                            b[ik] = new_ratio
+                    else:
+                        # Ratio not variable
+                        if br.ratio > br.ratio_max:
+                            br.ratio = br.ratio_max
+                        elif br.ratio < br.ratio_min:
+                            br.ratio = br.ratio_min
+                        else:
+                            # Find closest ratio
+                            dratio = (br.ratio_max - br.ratio_min) / (br.num_ratios-1)
+                            ptn_ratios = np.arange(0, br.num_ratios)*dratio + br.ratio_min
+                            i_close = np.argmin(np.abs(ptn_ratios - br.ratio))
+                            br.ratio = ptn_ratios[i_close]
+                        
         # Update
         solver.func(x)
         p.update_lin()