+++ CHANGELOG.md
@@ -2,6 +2,16 @@ Unreleased
 ----------
 * Added variables on Inter-Area Transfer.
 * Added constraints and heuristics on regulating inter-area transfer MW.
+* Added transformer P, Q control constraints and heuristics in NR.
+* Added transformer P, Q control functions in Opt-based ACPF.
+* Added phase shift varaibles for phase shifters.
+* Added asymmetric transformer equation on ratio and phase angle in Opt-based.
+* Added support for fixed power factor generator controls (e.g., wind machines).
+* Added support for Q limit regulation from generator power factors (e.g., wind machines).
+* Add FACTS constraints with for series voltage control and series impedance control.
+* Added `loads_2_ZIP` parameter for ACPF to convert loads to constant imepdance if voltage drops beloew threshold (default is False).
+* Adding admittance correction parameter `y_correction` and callback to ACPF Newton-Raphson only.
+* Added tap-changer rounding parameter `taps_round` that will round taps each step if tap-mode is regulating.
 
 Verions 1.4.1
 -------------